PatRec-3 add get lines endpoint

+++ src/main/java/com/example/demo/entity/Point.java
@@ -8,12 +8,12 @@ public class Point {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private float x;
-    private float y;
+    private double x;
+    private double y;
 
     public Point() {}
 
-    public Point(float x, float y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
@@ -26,19 +26,19 @@ public class Point {
         this.id = id;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
